pyproject.toml
[project]
name = "chatgpt2api"
version = "1.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.15.0",
    "fastapi>=0.136.0",
    "pillow>=12.2.0",
    "pybase64>=1.4.3",
    "python-multipart>=0.0.26",
    "tiktoken>=0.12.0",
    "uvicorn>=0.44.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "gitpython>=3.1.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
